Added Weakpoints
Weakpoints are subclasses of ActorHealth, and work almost the same way, except they also deal damage to their "owner" (the thing that uses this weakpoint as its weakpoint), equal to the damage taken times a predefined multiplier. They can also be configured to bypass damage resistances of the owner.

Of, and ActorHealth now has a damage resistance, on a scale of 0 (no resistance) to 1 (invincible)